// backend/src/common/enums.ts
export const TapSource = {
  NFC_UID: 'NFC_UID',
  NFC_NDEF: 'NFC_NDEF',
  QR: 'QR',
  MANUAL: 'MANUAL',
} as const;
export type TapSource = (typeof TapSource)[keyof typeof TapSource];